Added ability to create todos from the UI
