+++ libs/mapper/src/lib/profiles/todo-mapping-profile.ts
@@ -84,11 +84,11 @@ const todoMappingProfile: MappingProfile = (mapper) => {
         )
         .forMember(
             (dest) => dest.created_by_id,
-            mapFrom((source) => source.session.userId)
+            mapFrom((source) => source.session?.userId)
         )
         .forMember(
             (dest) => dest.assigned_to_id,
-            mapFrom((source) => source.assignedToId ?? source.session.userId)
+            mapFrom((source) => source.assignedToId ?? source.session?.userId)
         );
 
     mapper
